Add clarification for sort order
Adding some clarifying language about how the sort works, specifically about how to change the order to ascending when presenting results.

According to this line [1], the sort order is only looking for `asc` and just assumes descending otherwise. I wrote the docs in a way to be explicit with the descending order, only to make it easier to understand. If you prefer we only make a documentation change that explicitly reflects the `asc` option, I'm happy to refactor.


[1]: https://github.com/omnivore-app/omnivore/blob/1cabdde6797e37da7f44e726b83d3f95d8b6bbcc/packages/api/src/utils/search.ts#L189